udev: add recipe to build from git (currently 168 + handfull of commits)
This recipe removes all but one of the custom rules OE installs. The local.rules will need some looking at, since it assumes ifupdown support while the image might be use connman/networkmanager/etc instead.

The compat-wrapper and machine specific rules where also removed. For udev 169 and beyond the machine specific rules should be in their own recipes or sub-package that is in udev RRECOMMENDS. This ensures that ${PN} is never machine specific

An empty udev-cache.service is installed to tell systemd to skip this sysV script

Tested on: beagleboard/angstrom

Signed-off-by: Koen Kooi <koen@openembedded.org>