aurora: use qte.sh script at bootup to get the correct touchscreen configuration
After the boot the /etc/profile.d directory will only be evaluated when the user logs in
so the QWS_MOUSE_PROTO environment variable is not set when the aurora daemon is started
automatically by the init system. To workaround this we include the /etc/profile.d/qte.sh
provided by qt4-embedded which is needed to execute aurora.

Signed-off-by: Simon Busch <morphis@gravedo.de>