initscripts: Modify 'populate-volatile.sh' to atomically update 'volatile.cache'
When power is lost before a new volatile.cache has been flushed to disk, it can
be corrupted, leaving the volatiles uncreated and causing software failure
during the next reboot.

Since only the existence of volative.cache is checked for using a temporary file
on the same volume and renaming it avoids this issue.

The file is sync'd before the rename to avoid problems with newer caching
filesystems, in which the rename is atomic, but the contents of the file may not
yet be flushed.

Signed-off-by: James Limbouris <james@digitalmatter.com.au>
Acked-by: Paul Menzel <paulepanter@users.sourceforge.net>