crimsonfields: update to 0.5.3
In the following version, do_stage was removed.

Native BBCLASSEXTEND was not used,since
crimsonfields and crimsonfields-native share very little.

Note that the previous native crimsonfields-native didn't configure anymore:
SDL_CONFIG pointed to "${STAGING_BINDIR}/sdl-config-native" which
  which didn't exist:
    checking for SDL - version >= 1.1.5...
    .../work/x86_64-linux/crimsonfields-native-0.4.8-r0/crimson-0.4.8/configure:
    line 3568: .../sysroots/x86_64-linux/usr/bin/sdl-config-native: No such file or directory

The resulting package was tested on the nokia900 target under the shr distribution.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@no-log.org>