libshiboken: Do not override `DEPENDS`
This is a fix up for commit 1b9260e7 [1].

	commit 1b9260e7a6c756d6938821f013a41c371065980a
	Author: Paul Menzel <paulepanter@users.sourceforge.net>
	Date:   Sun Jun 19 11:45:37 2011 +0200

		  shiboken: Add `apiextrator-native` to `DEPENDS`

This commit does not fix the mentioned error, that `ApiExtractorConfig.cmake` is not built, since by using `=` instead of `+=` the dependencies from the include file `shiboken.inc` are overriden in the recipe.

This patch fixes the error and is built tested using `minimal{,-uclibc}` for `MACHINE = "beagleboard"`.

[1] http://cgit.openembedded.org/cgit.cgi/openembedded/commit/?id=1b9260e7a6c756d6938821f013a41c371065980a

Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>
Acked-by: Simon Busch <morphis@gravedo.de>