frameworkd: check even in ${MACHINE_CLASS} directory for a config file and bump PR
This fixes a problem with the palmpre2 machine which does not get the right
frameworkd.conf configuration file as the recipes searched in a directory with ${MACHINE}
as name which is not there for the palmpre2. So it tooks the general one which is wrong.
With checking also in the ${MACHINE_CLASS} directory for the configuration
(${MACHINE_CLASS} is palmpre for both palmpre and palmpre2) it now takes the right
configuration file.

Signed-off-by: Simon Busch <morphis@gravedo.de>