htop: Link with libubacktrace when compiling for uClibc
Using `minimal-uclibc` for `MACHINE = "beagleboard"` task compile fails with the following error.

	htop-CRT.o: In function `CRT_handleSIGSEGV':
	/oe/build-minimal-uclibc/minimal-uclibc-dev/work/armv7a-oe-linux-uclibceabi/htop-0.9-r0/htop-0.9/CRT.c:131: undefined reference to `backtrace'
	/oe/build-minimal-uclibc/minimal-uclibc-dev/work/armv7a-oe-linux-uclibceabi/htop-0.9-r0/htop-0.9/CRT.c:133: undefined reference to `backtrace_symbols_fd'

Backtrace support in uClibc was enabled in commit 2a1c9e1e [1] so appending `-lubacktrace` to `LDFLAGS` fixes this problem.

This patch is ported from the meta-openembedded patch committed from Khem Raj in commit 7e3451f4 [2].

PR was not increased, because it did not compile before with uClibc and the package content does not change.

[1] http://cgit.openembedded.org/cgit.cgi/openembedded/commit/?id=2a1c9e1ee5eb8067592d26100b563dd2ae7be00d
[2] http://cgit.openembedded.org/cgit.cgi/meta-openembedded/commit/?id=7e3451f4e7352cc2c48711170e6a1be7c15e973e

Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>