mplayer_git: Add patch to fix linker error on x86 machines
For x86 machines the option `--enable-runtime-cpudetection` is passed to the configure script.

        EXTRA_OECONF_append = " ${@base_contains('MACHINE_FEATURES', 'x86', '--enable-runtime-cpudetection', '',d)} "

This causes a linker error like

        fix undefined references  in loader/module.o:
        module.c:(.text+0xd60): undefined reference to `report_entry'
        module.c:(.text+0xd6a): undefined reference to `report_ret'
        module.c:(.text+0xd74): undefined reference to `wrapper_target'
        module.c:(.text+0xd7a): undefined reference to `wrapper'

as reported and fixed in the Gentoo’s Bugzilla Bug 213836 [1]. The patch was partly adopted and applied already to `mplayer_svn` in OpenEmbedded in commit »mplayer_svn : fix compilation problem« (ff57f14c) [2].

This patch is build tested for `DISTRO = "shr"` with a modified `conf/machine/ion.conf`.

[1] https://bugs.gentoo.org/213836?id=213836
[2] http://cgit.openembedded.org/cgit.cgi/openembedded/commit/?id=ff57f14c99a6ff61ffb7f1c533523e3e934f96e5

Signed-off-by: Josua Mayer <josua.mayer97@googlemail.com>
Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>