classpath in 2011.03-maintenance
On Fri, 2011-08-05 at 11:38 -0400, George C. Huntington, III wrote:
> I posted a patch that makes the wrapper script return the error code
> if it
> fails after 5 retries.  it is on patchwork.
> http://patches.openembedded.org/patch/7249/

It seems like saving the javac script arguments into the ARGS variable
is what broke the old behavior where -bootclasspath '' worked for
building classpath.  Putting that part of the script back to the way it
used to be fixes it.

Do you want to merge this in with your error code return patch?

Otherwise, I can submit it as another patch, but it will depend on yours...

Thanks,
Jesse