gcc: Remove package libstdc++-dev and libgcc-dev from gcc-cross
Remove package libstdc++-dev and libgcc-dev from gcc-cross as they
already provided by gcc.

Commit 8c42351519ece9ed60659a103e624d2e154b3641 add dev packages
(libstdc++-dev libgcc-dev) to gcc-cross (gcc-package-cross.inc)
which are already part of gcc. The libgcc-dev package both include
the libgcc_s.so file which triggers a 'file is already provided by
package' error on incremental builds. The libstdc++-dev package have
Acked-by: Steffen Sledz <sledz@dresearch-fe.de>

different content and is empty on some archs (armv5te, armv7e).
The empty package triggers a 'Cannot satisfy the following
dependencies for task-sdk-bare' error. If an empty libstdc++-dev
package is generated it triggers a '..../usr/include/c++ not exist'
error as this is part of the package from gcc but not gcc-cross.

Signed-off-by: Stefan Herbrechtsmeier <sherbrec@cit-ec.uni-bielefeld.de>
Acked-by: Steffen Sledz <sledz@dresearch-fe.de>