binutils: Update from 2.21 to 2.21.1a
Due to a GPL violation corrected Binutils archives were uploaded with the suffix »a« and the faulty archives were removed [1]. The URLs redirect to the new archives and therefore the checksums are now incorrect.

Use this incident to update the package also to the new version 2.21.1a. Due to this one patch had to be regenerated and was taken from OE-core [2].

This patch is build tested using `angstrom-2010.x` and `minimal{,-uclibc}` for `MACHINE = "beagleboard"`.

[1] http://nickclifton.livejournal.com/9067.html
[2] http://git.openembedded.org/cgit.cgi/openembedded-core/tree/meta/recipes-devtools/binutils/binutils/libtool-2.4-update.patch?id=4810404f22d7cf8849dc32da68dd33a156cac407

Signed-off-by: Sylvain 'GarthPS' Paré <sylvain.pare@gmail.com>
Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>