u-boot: added 2011.03; also added building of tools
In Aug 2010 the building of u-boot utils (fw_printenv, fw_setenv) changed
Since then they are build using HOSTCC, which does not really give
a good target image.

After discussion with Wolfgang Denk and backreading the u-boot mailing
list, I opted to redefine HOSTCC in the recipe when building env
That way we can get u-boot tools on the target
(btw,  there is a recipe u-boot-utils_1.2.0.bb but that uses
a highly outdated version of u-boot mkimage that does not support
FIT  (Flattended Image Tree) images)

(FIT image are quite useful on powerpc if one wants to
build a blob that contains both a uimage and a device tree)

I've been using this version of u-boot on mpc8313 for more than half
a year and FIT images for about 4 months.

BTW: wrt the patch to fw_env.h:
The ifndefs are needed to actually compile the tools.
The changed constants are values that are more common nowadays
(16 k flash pages is probably not too common; 128K is way more common)
(one probably wants to change this for a specific machine anyway).

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>